Change algorithm for text rotater
The original algorithm was a typer, and while looking nice, I have seen
it before on other sites.

The new algorithm is based on the levenshtein distance between two
strings. In the algorithm to calculate the minimum edits, we can
also calculate the edits that need to be made to the strings. These
edits will appear above and below the current text and then combine
with the text to form the next string.