Update edit distance algorithm
The algorithm still has the same complexity, however instead of building
the dp array and then traversing backwards to obtain the move sequence,
the move sequence is calculated on the fly.

This ensures that the number of edits is at the minimum.

There will be no threat of a memory leak if the garbage collector will
collect unreferenced variables.