Change width and height style
I want to app to fill the page, so I previously had 100vw and 100vh for
the width and height of the App container respectively.

This is a bit of a hack however, as the view height is more than 100% on
mobile browsers like chrome because of the retractable address bar.

I have changed it to position fixed and width and height both 100%.